// packages/editor/lib/models/annotation.model.ts
export interface HandleKeyDownDeps {
  clipboard?: { objects: unknown[] };
  e: KeyboardEvent;
  canvas: Canvas;
  undo: () => void;
  redo: () => void;
  syncShapeInStorage: (shape: FabricObject) => void;
  deleteShapeFromStorage: (id: string) => void;
}
// ...
export interface CustomFabricObject<T extends FabricObject> extends FabricObject {
  objectId?: string;
}